Move OperationList in its own module (to later fix circular dependency issue)
